Fix incompatibility with Cython 3.0.0a10
In this alpha release of Cython 3, the behavior of
`Cython.Distutils.build_ext` has changed, and now the `distutils: sources`
directive in `_mask.pyx` is interpreted. This means that `maskApi.c` is
added to the extension twice (first by `setup.py`, second by `_mask.pyx`),
so it's built twice and then linking fails due to multiple symbol
definitions.

If the project is built from an sdist, the reference to
`../common/maskApi.c` refers to a nonexistent file (note: the path is
interpreted relative to the project, not relative to `_mask.pyx`), so the
build also fails, but in a different way.

Remove the redundant directive to fix this.